Added url input for textures (#5)
#2

This was harder than expected, right now it expects a callback, the way I implemented it though doesn't need one. Going to remove it. For now I have a simple cache, might move it to the filesystem, or remove it all together.

#4
Added a callback that can be set at time of tab sprite population to alter the sprite. Example in the example code.